+++ He/Tasks/Task1/app.css
@@ -1,0 +1,5 @@
+* {
+    margin: 0;
+    padding: 0;
+    list-style:none;
+}
