+++ He/Tasks/Task1/example/src/.jshintrc
