+++ He/Tasks/Task1/src/app.js
@@ -1,22 +1,12 @@
+// import {
+//     Store,$on
+// } from './util'
+// import Controller from './controller'
+// import View from './view'
 
-import Controller from './controller';
-import {$on} from './helpers';
-import Template from './template';
-import Store from './store';
-import View from './view';
+// const view = new View()
+// const store = new Store('myTodo')
+// const controller = new Controller(view,store)
 
-const store = new Store('text');
-
-const template = new Template();
-const view = new View(template);
-
-/**
- * @type {Controller}
- */
-const controller = new Controller(store, view);
-
-const setView = () => controller.setView(document.location.hash);
-$on(window, 'load', setView);
-$on(window, 'hashchange', setView);
-
-
+// $on(window,'load')
+// $on(window,'hashChange')
